requirements.txt
paramiko==2.4.1
wakeonlan==1.1.6
python_nmap==0.6.1
cement==2.10.2
nmap==0.0.1
PyYAML==3.12
